added chaos-y functional test
