+++ tests/functional/utils.py
@@ -1,0 +1,222 @@
+# Copyright 2016 Workiva Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+# system imports
+import base64
+import json
+import unittest
+import random
+
+# library imports
+from botocore.exceptions import ClientError
+
+# application imports
+from aws_lambda_fsm.client import start_state_machine
+from aws_lambda_fsm import handler
+
+
+class Messages(object):
+
+    def __init__(self):
+        self.messages = []
+        self.all_messages = []
+
+    def send(self, data):
+        self.messages.append(data)
+        self.all_messages.append(data)
+
+    def recv(self):
+        return self.messages.pop(0) if self.messages else None
+
+    def reset(self):
+        del self.messages[:]
+        del self.all_messages[:]
+
+    def trace(self, uvars=(), raw=False):
+        s = 'system_context'
+        u = 'user_context'
+        svars = ('current_state', 'current_event', 'steps', 'retries')
+        serialized = map(lambda x: json.loads(x), self.all_messages)
+        if raw:
+            return serialized
+        data = enumerate(serialized)
+        return map(
+            lambda x: (
+                x[0],
+                tuple(x[1][s][v] for v in svars),
+                tuple(x[1][u].get(v) for v in uvars)
+            ),
+            data
+        )
+
+
+# not threadsafe, yada yada
+class AWSStub(object):
+
+    def __init__(self):
+        self.primary_stream_source = Messages()
+        self.secondary_stream_source = Messages()
+
+        self.primary_retry_source = Messages()
+        self.secondary_retry_source = Messages()
+
+        self.all_sources = Messages()
+
+        self.primary_cache = {}
+        self.secondary_cache = {}
+
+        self.all_caches = {}
+
+        self.primary_stream_source_failure = False
+        self.primary_retry_source_failure = False
+
+        self.primary_chaos = 0.0
+        self.secondary_chaos = 0.0
+
+        self.errors = Messages()
+
+    def _chaos(self, primary):
+        chaos = {True: self.primary_chaos, False: self.secondary_chaos}[primary]
+        if chaos > 0.0:
+            if random.uniform(0.0, 1.0) < chaos:
+                raise ClientError({"Error": {}}, "chaos")
+
+    def _lease_chaos(self, primary):
+        # lease acquisition has own chaos
+        chaos = {True: self.primary_chaos, False: self.secondary_chaos}[primary]
+        if chaos > 0.0:
+            if random.uniform(0.0, 1.0) < chaos:
+                return False
+
+    def _get_stream_source(self, primary):
+        self._chaos(primary)
+        if self.primary_stream_source_failure and primary:
+            raise ClientError({"Error": {}}, "mock")
+        return {
+            True: self.primary_stream_source,
+            False: self.secondary_stream_source
+        }[primary]
+
+    def _get_cache_source(self, primary):
+        return {
+            True: self.primary_cache,
+            False: self.secondary_cache
+        }[primary]
+
+    def _get_retry_source(self, primary):
+        self._chaos(primary)
+        if self.primary_retry_source_failure and primary:
+            raise ClientError({"Error": {}}, "mock")
+        return {
+            True: self.primary_retry_source,
+            False: self.secondary_retry_source
+        }[primary]
+
+    def get_message(self):
+        return self.primary_stream_source.recv() or \
+            self.secondary_stream_source.recv() or \
+            self.primary_retry_source.recv() or \
+            self.secondary_retry_source.recv()
+
+    def send_next_event_for_dispatch(self, context, data, correlation_id, delay=0, primary=True):
+        self._get_stream_source(primary).send(data)
+        self.all_sources.send(data)
+        return {'test': 'stub'}
+
+    def set_message_dispatched(self, correlation_id, steps, retries, primary=True):
+        key = '%s-%s' % (correlation_id, steps)
+        self._get_cache_source(primary)[key] = True
+        self.all_caches[key] = True
+        return True
+
+    def get_message_dispatched(self, correlation_id, steps, primary=True):
+        self._get_cache_source(primary).get('%s-%s' % (correlation_id, steps))
+
+    def acquire_lease(self, correlation_id, steps, retries, primary=True):
+        self._lease_chaos(primary)
+        key = 'lease-%s-%s' % (correlation_id, steps)
+        if key in self._get_cache_source(primary):
+            return False
+        else:
+            self._get_cache_source(primary)[key] = True
+            self.all_caches[key] = True
+            return True
+
+    def release_lease(self, correlation_id, steps, retries, fence_token, primary=True):
+        self._lease_chaos(primary)
+        key = 'lease-%s-%s' % (correlation_id, steps)
+        if key in self._get_cache_source(primary):
+            self._get_cache_source(primary).pop(key)
+            return True
+        else:
+            return False
+
+    def start_retries(self, context, run_at, payload, primary=True):
+        self._get_retry_source(primary).send(payload)
+        self.all_sources.send(payload)
+        return {'test': 'stub'}
+
+    def increment_error_counters(self, data, dimensions):
+        self.errors.send(json.dumps((data, dimensions)))
+        return {'test': 'stub'}
+
+    def store_checkpoint(self, context, sent, primary=True):
+        return {'test': 'stub'}
+
+    def reset(self):
+        self.primary_stream_source.reset()
+        self.secondary_stream_source.reset()
+        self.primary_retry_source.reset()
+        self.secondary_retry_source.reset()
+        self.all_sources.reset()
+        self.primary_cache.clear()
+        self.secondary_cache.clear()
+        self.all_caches.clear()
+        self.primary_stream_source_failure = False
+        self.primary_retry_source_failure = False
+        self.errors.reset()
+        self.primary_chaos = 0.0
+        self.secondary_chaos = 0.0
+
+
+def to_kinesis_message(data):
+    return {
+        "Records": [{
+            "kinesis": {
+                "data": base64.b64encode(data)
+            }
+        }]
+    }
+
+
+class BaseFunctionalTest(unittest.TestCase):
+
+    def _execute(self, aws, machine_name, context,
+                 fail_primary_stream_stream=False,
+                 fail_primary_stream_retry=False,
+                 primary_chaos=0.0,
+                 secondary_chaos=0.0):
+        start_state_machine(machine_name, context, correlation_id='correlation_id')
+        if fail_primary_stream_stream:
+            aws.primary_stream_source_failure = True
+        if fail_primary_stream_retry:
+            aws.primary_retry_source_failure = True
+        if primary_chaos:
+            aws.primary_chaos = primary_chaos
+        if secondary_chaos:
+            aws.secondary_chaos = secondary_chaos
+        message = aws.get_message()
+        while message:
+            handler.lambda_kinesis_handler(to_kinesis_message(message))
+            message = aws.get_message()
